+++ src/algorithms/constructors/mod.rs
@@ -8,6 +8,7 @@
 pub mod atlas;
 pub mod atlas_edges;
 pub mod circulant;
+pub mod create;
 pub mod de_bruijn;
 pub mod famous;
 pub mod full;